backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn[standard]==0.24.0
pyjwt==2.8.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2